refactor(reborn): extract WASM blocking-execution machinery to its own module
Pure move, no behavior change. runtime_adapters.rs had grown to mix
adapter dispatch selection with the reservation RAII state machine, the
blocking-pool semaphore gates, and the spawn_blocking bridge for WASM
prepare/execute — too many ownership concerns for one file.

Move the semaphore gates, ReservationGuard, execute_prepared_wasm,
run_wasm_execution_blocking, run_wasm_prepare_blocking, the WASM-only
error/guest helpers, and their tests into a new sibling module
services/wasm_execution.rs. ReservationGuard and execute_prepared_wasm
are pub(super) since the first-party and WASM adapters (which stay in
runtime_adapters.rs) still use them; wasm_error_kind stays put and is
imported back.

runtime_adapters.rs: 2089 -> 708 lines. No logic, signatures, or
messages changed; clippy clean, full lib suite green.
